cast an unsigned to a double before calling log() - Intel compiler issues

+++ Signal/General/Response.C
@@ -102,7 +102,7 @@ unsigned dsp::Response::get_minimum_ndat () const
   if (nsmear == 0)
     return 0;
 
-  return (unsigned) pow (2.0, 1.0 + ceil( log(nsmear)/log(2.0) ));
+  return (unsigned) pow (2.0, 1.0 + ceil( log((double)nsmear)/log(2.0) ));
 }
 
 /*!  Using the get_minimum_ndat method and the max_ndat static attribute,
